requirements.txt
pandas
scikit-learn
numpy
seaborn
Flask
#this is more stable version
mlflow==2.2.2
dvc
cassandra-driver

-e .